requirements.txt
Flask==2.1.2
Pillow==9.2.0
scikit_image==0.19.2
skimage==0.0
torch==1.8.2+cu102
torchvision==0.9.2+cu102
